Work around S3 object names with leading slashes.
It seems that the default prefix for Cost & Usage Reports starts with a
slash, which results in S3 urls like
https://bucket.s3.amazonaws.com//reportname/date-range/reportname-00001.csv.gz.
This extra / was causing s3cmd to attempt to copy/hardlink the files to
/reportname/... on the local filesystem (which failed due to lack of
permissions on /), and the absence of the files in
that location was causing it to always download them.  This workaround
strips leading slashes from src file names.